# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/warehouses": {
      "get": {
        "operationId": "findAllWarehouses",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for warehouses",
            "schema": {
              "example": "Main",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Warehouses retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedWarehouseResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all warehouses",
        "tags": [
          "Warehouses"
        ]
      },
      "post": {
        "operationId": "createWarehouse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWarehouseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Warehouse created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new warehouse",
        "tags": [
          "Warehouses"
        ]
      }
    },
    "/warehouses/{id}": {
      "get": {
        "operationId": "findOneWarehouse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Warehouse retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Warehouse not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific warehouse by id",
        "tags": [
          "Warehouses"
        ]
      },
      "patch": {
        "operationId": "updateWarehouse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarehouseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Warehouse updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "404": {
            "description": "Warehouse not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a specific warehouse by id",
        "tags": [
          "Warehouses"
        ]
      },
      "delete": {
        "operationId": "removeWarehouse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Warehouse deleted successfully."
          },
          "404": {
            "description": "Warehouse not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a specific warehouse by id",
        "tags": [
          "Warehouses"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "The API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "WarehouseResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the warehouse"
          },
          "name": {
            "type": "string",
            "description": "The name of the warehouse"
          },
          "description": {
            "type": "string",
            "description": "The description of the warehouse"
          },
          "address": {
            "type": "string",
            "description": "The address of the warehouse"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "The last update date"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedWarehouseMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalItems": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "page",
          "limit",
          "totalItems",
          "totalPages"
        ]
      },
      "PaginatedWarehouseResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WarehouseResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedWarehouseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateWarehouseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Main Logistics Center",
            "description": "The name of the warehouse"
          },
          "description": {
            "type": "string",
            "example": "Located in downtown area.",
            "description": "The description of the warehouse"
          },
          "address": {
            "type": "string",
            "example": "123 Warehouse St, City",
            "description": "The address of the warehouse"
          }
        },
        "required": [
          "name",
          "address"
        ]
      },
      "UpdateWarehouseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Main Logistics Center",
            "description": "The name of the warehouse"
          },
          "description": {
            "type": "string",
            "example": "Located in downtown area.",
            "description": "The description of the warehouse"
          },
          "address": {
            "type": "string",
            "example": "123 Warehouse St, City",
            "description": "The address of the warehouse"
          }
        }
      },
...
    }
  }
}